refactor: reorganize imports and update file paths for PostCard and PostSkeleton components delete: remove unused authentication forms and schemas fix: adjust queryClient configuration for clarity
